Made changes in how controllers are called by their respective models (now using invoke method) and how controllers pass values to views. Made slight visual changes on the form to make everything bigger and easier to see.
Login system created (but no cookies yet). On pressing Submit
button, page should request for itself (in this case, page is called
home, but  will do since .htaccess will convert it into home).
App will calll Request to figure out which controller to call.
Once HomeController is invoked, it will check if username and
password are set and if so will call HomeModel to handle login.
HomeModel will check login data, request data from DatabaseHandler
and decide to return a message to HomeController if login is unvalid,
or it will redirect the page if login is successful. HomeController
finally passes messages to view and the page is displayed with
feedback from login attempt.